feat(rdf-core): the cover law moves to the kernel, leaner than it left
Every ordered codec will need the same decode half, so it lives where
the verification law already lives: purrdf_core::cover owns
reconstruct(), VerbatimSpan and ReconstructError, format-neutral, and
purrdf-markdown re-exports them as its first emitter. The next
structured format costs an emitter and nothing else, and two codecs
cannot drift apart on what a lawful cover is — SPEC §11.3 now states
that in §5.1's own words.

The move carries an allocation audit with it. The declared-piece
lookup binary-searches the sorted span vector instead of building a
map nothing needed on an overlap-free cover; the output buffer is
append-only at exact capacity — no zero-fill pass, no byte written
twice where a continuation overlaps; overlap agreement is one slice
comparison, and only a refusal pays to locate the differing byte; and
the one u64-to-usize cast states its proof through try_from instead of
truncating silently on a 32-bit target.
